backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# NLP & NER
spacy==3.7.2
rapidfuzz==3.6.1

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Security & Encryption
cryptography==42.0.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
groq==0.4.2

# Rate Limiting
slowapi==0.1.9

# Utilities
uuid==1.30
